Split Lexer into it's own header; added mode to skip EndOfLine tokens
