+++ PlaceRouteHierFlow/PnRDB/unit_tests.cpp
@@ -49,6 +49,33 @@ TEST(PnRTest, Lexer) {
   }
 }
 
+TEST(PnRTest, Lexer2) {
+  string str =
+"//Verilog block level netlist file for telescopic_ota\n"
+"//Generated by UMN for ALIGN project\n" 
+"\n"
+"\n"
+"module telescopic_ota ( voutn, vss, vinp, vbiasp1, vbiasnd, vbiasn, voutp, vbiasp2, vinn, d1, vdd ); \n"
+"input voutn, vss, vinp, vbiasp1, vbiasnd, vbiasn, voutp, vbiasp2, vinn, d1, vdd; \n"
+"\n"
+"CMC_PMOS_S_n12_X1_Y1 m2_m1 ( .DA(net012), .G(vbiasp1), .DB(net06), .S(vdd) ); \n"
+"DP_NMOS_n12_X3_Y3 m3_m0 ( .DA(net014), .GA(vinn), .S(net10), .DB(net8), .GB(vinp) ); \n"
+"CMFB_NMOS_n12_X3_Y1 m5_m4 ( .DA(d1), .S(vss), .DB(net10), .GB(vbiasnd) ); \n"
+"CMC_PMOS_n12_X2_Y1 m6_m7 ( .DA(voutn), .G(vbiasp2), .DB(voutp), .SA(net06), .SB(net012) ); \n"
+"CMC_NMOS_n12_X3_Y1 m9_m8 ( .DA(voutn), .G(vbiasn), .DB(voutp), .SA(net8), .SB(net014) ); \n"
+"\n"
+"endmodule\n";
+
+  std::istringstream is(str);
+  Lexer l(is,1);
+
+  while ( !l.have( TokenType::EndOfFile)) {
+    std::cout << "(" << l.current_token.tt << ",\"" << l.current_token.value << "\")" << std::endl;
+    EXPECT_NE( l.current_token.tt, TokenType::EndOfLine);
+    l.get_token();
+  }
+}
+
 TEST(PnRTest, ReadVerilog) {
   string str =
 "//Verilog block level netlist file for telescopic_ota\n"
